# src/main/resources/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "PowerCards Support",
      "url" : "https://github.com/victobonetti/powercards"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "PowerCards API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/v1/tags" : {
      "get" : {
        "summary" : "List tags",
        "parameters" : [ {
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "Tag Resource" ]
      },
      "post" : {
        "summary" : "Create a new tag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tag"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Tag Resource" ]
      }
    },
    "/v1/tags/stats" : {
      "get" : {
        "summary" : "List tags with stats",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "perPage",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedResponseTagStats"
                }
              }
            }
          }
        },
        "tags" : [ "Tag Resource" ]
      }
    },
    "/v1/tags/{id}" : {
      "delete" : {
        "summary" : "Delete a tag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "Tag Resource" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "PaginatedResponseTagStats" : {
        "type" : "object",
        "properties" : {
          "pagination" : {
            "$ref" : "#/components/schemas/PaginationMeta"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TagStats"
            }
          }
        }
      },
      "PaginationMeta" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nextPageUri" : {
            "type" : "string"
          },
          "lastPageUri" : {
            "type" : "string"
          }
        }
      },
      "Tag" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "TagStats" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "noteCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8088",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8088",
    "description" : "Auto generated value"
  } ]
}